Refactor chat page: split components, improve tool loading, and enhance UI
