+++ chat_page.dart
@@ -4,15 +4,16 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:http/http.dart' as http;
 import 'package:image_picker/image_picker.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'models.dart';
 import 'character_service.dart';
 import 'external_tools_service.dart';
+import 'chat_message_bubble.dart';
+import 'chat_input_bar.dart';
+import 'chat_panels.dart';
+import 'tool_loading_indicator.dart';
 
 /* ----------------------------------------------------------
    CHAT PAGE
@@ -898,7 +899,7 @@ Error: ${result['error']}''';
               itemCount: _messages.length,
               itemBuilder: (_, index) {
                 final message = _messages[index];
-                return _MessageBubble(
+                return MessageBubble(
                   message: message,
                   onRegenerate: () => _regenerateResponse(index),
                   onUserMessageTap: () => _showUserMessageOptions(context, message),
@@ -946,76 +947,13 @@ Error: ${result['error']}''';
                 ],
               ),
             ),
-          // Enhanced external tools status with better design
-          if (_externalToolsService.isExecuting)
-            Container(
-              margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                color: const Color(0xFFEAE9E5),
-                borderRadius: BorderRadius.circular(16),
-                border: Border.all(color: const Color(0xFFD1D1D1), width: 1),
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Row(
-                    children: [
-                      Container(
-                        width: 20,
-                        height: 20,
-                        child: const CircularProgressIndicator(
-                          strokeWidth: 2,
-                          color: Color(0xFF000000),
-                        ),
-                      ),
-                      const SizedBox(width: 12),
-                      Expanded(
-                        child: Text(
-                          _externalToolsService.currentlyExecutingTools.length > 1
-                              ? 'Executing ${_externalToolsService.currentlyExecutingTools.length} tools in parallel'
-                              : 'Using external tool',
-                          style: const TextStyle(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                            color: Color(0xFF000000),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  if (_externalToolsService.currentlyExecutingTools.isNotEmpty) ...[
-                    const SizedBox(height: 8),
-                    Wrap(
-                      spacing: 8,
-                      runSpacing: 4,
-                      children: _externalToolsService.currentlyExecutingTools.map((tool) {
-                        return Container(
-                          padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                          decoration: BoxDecoration(
-                            color: const Color(0xFF000000),
-                            borderRadius: BorderRadius.circular(12),
-                          ),
-                          child: Text(
-                            tool,
-                            style: const TextStyle(
-                              fontSize: 12,
-                              color: Color(0xFFFFFFFF),
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                        );
-                      }).toList(),
-                    ),
-                  ],
-                ],
-              ),
-            ),
+          // Enhanced external tools status with better animation
+          ToolLoadingIndicator(externalToolsService: _externalToolsService),
                         SafeArea(
             top: false,
             left: false,
             right: false,
-            child: _InputBar(
+            child: ChatInputBar(
               controller: _controller,
               onSend: () => _send(),
               onStop: _stopGeneration,
@@ -1037,1482 +975,15 @@ Error: ${result['error']}''';
 }
 
 /* ----------------------------------------------------------
-   MESSAGE BUBBLE & ACTION BUTTONS - iOS Style Interactions
+   CHAT PAGE IMPLEMENTATION
 ---------------------------------------------------------- */
-class _MessageBubble extends StatefulWidget {
-  final Message message;
-  final VoidCallback? onRegenerate;
-  final VoidCallback? onUserMessageTap;
-  const _MessageBubble({
-    required this.message,
-    this.onRegenerate,
-    this.onUserMessageTap,
-  });
-
-  @override
-  State<_MessageBubble> createState() => _MessageBubbleState();
-}
-
-class _MessageBubbleState extends State<_MessageBubble> with TickerProviderStateMixin {
-  bool _showActions = false;
-  late AnimationController _actionsAnimationController;
-  late Animation<double> _actionsAnimation;
-  bool _showUserActions = false;
-  late AnimationController _userActionsAnimationController;
-  late Animation<double> _userActionsAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _actionsAnimationController = AnimationController(
-      duration: const Duration(milliseconds: 200),
-      vsync: this,
-    );
-    _actionsAnimation = CurvedAnimation(
-      parent: _actionsAnimationController,
-      curve: Curves.easeOut,
-    );
-    
-    _userActionsAnimationController = AnimationController(
-      duration: const Duration(milliseconds: 200),
-      vsync: this,
-    );
-    _userActionsAnimation = CurvedAnimation(
-      parent: _userActionsAnimationController,
-      curve: Curves.easeOut,
-    );
-  }
-
-  @override
-  void dispose() {
-    _actionsAnimationController.dispose();
-    _userActionsAnimationController.dispose();
-    super.dispose();
-  }
-
-  void _toggleActions() {
-    setState(() {
-      _showActions = !_showActions;
-      if (_showActions) {
-        _actionsAnimationController.forward();
-      } else {
-        _actionsAnimationController.reverse();
-      }
-    });
-  }
-
-  void _toggleUserActions() {
-    setState(() {
-      _showUserActions = !_showUserActions;
-      if (_showUserActions) {
-        _userActionsAnimationController.forward();
-      } else {
-        _userActionsAnimationController.reverse();
-      }
-    });
-  }
-
-  void _giveFeedback(BuildContext context, bool isPositive) {
-    HapticFeedback.lightImpact();
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(
-          isPositive ? '👍 Thank you for your feedback!' : '👎 Feedback noted. We\'ll improve!',
-          style: const TextStyle(
-            color: Color(0xFF000000),
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-        margin: const EdgeInsets.all(16),
-        elevation: 4,
-        duration: const Duration(seconds: 2),
-      ),
-    );
-    // Hide actions after interaction
-    _toggleActions();
-  }
-
-  void _copyMessage(BuildContext context) {
-    HapticFeedback.lightImpact();
-    Clipboard.setData(ClipboardData(text: widget.message.text));
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: const Text(
-          '📋 Message copied to clipboard!',
-          style: TextStyle(
-            color: Color(0xFF000000),
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-        margin: const EdgeInsets.all(16),
-        elevation: 4,
-        duration: const Duration(seconds: 2),
-      ),
-    );
-    // Hide actions after interaction
-    _toggleActions();
-  }
-
-  void _shareMessage(BuildContext context) {
-    HapticFeedback.lightImpact();
-    // For now, copy to clipboard (can implement actual sharing later)
-    Clipboard.setData(ClipboardData(text: widget.message.text));
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: const Text(
-          '🔗 Message ready to share!',
-          style: TextStyle(
-            color: Color(0xFF000000),
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-        margin: const EdgeInsets.all(16),
-        elevation: 4,
-        duration: const Duration(seconds: 2),
-      ),
-    );
-    // Hide actions after interaction
-    _toggleActions();
-  }
-
-
-  Widget _buildImageWidget(String url) {
-    try {
-      Widget image;
-      if (url.startsWith('data:image')) {
-        final commaIndex = url.indexOf(',');
-        final header = url.substring(5, commaIndex);
-        final mime = header.split(';').first;
-        final base64Data = url.substring(commaIndex + 1);
-        final bytes = base64Decode(base64Data);
-        if (mime == 'image/svg+xml') {
-          image = SvgPicture.memory(bytes, fit: BoxFit.contain);
-        } else {
-          image = Image.memory(bytes, fit: BoxFit.contain);
-        }
-      } else {
-        if (url.toLowerCase().endsWith('.svg')) {
-          image = SvgPicture.network(
-            url,
-            fit: BoxFit.contain,
-            placeholderBuilder: (context) =>
-                const Center(child: CircularProgressIndicator()),
-          );
-        } else {
-          image = Image.network(url, fit: BoxFit.contain);
-        }
-      }
-
-      return ClipRRect(
-        borderRadius: BorderRadius.circular(12),
-        child: ConstrainedBox(
-          constraints: const BoxConstraints(maxHeight: 300, maxWidth: double.infinity),
-          child: image,
-        ),
-      );
-    } catch (_) {
-      return Container(
-        height: 200,
-        alignment: Alignment.center,
-        child: const Icon(Icons.image_not_supported, color: Colors.grey),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isBot = widget.message.sender == Sender.bot;
-    final isUser = widget.message.sender == Sender.user;
-    final canShowActions = isBot && !widget.message.isStreaming && widget.message.text.isNotEmpty && widget.onRegenerate != null;
-
-    Widget bubbleContent = Container(
-      margin: const EdgeInsets.symmetric(vertical: 8),
-      padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 20),
-      constraints: const BoxConstraints(maxWidth: 320),
-      decoration: BoxDecoration(
-        color: isBot ? Colors.transparent : const Color(0xFFEAE9E5),
-        borderRadius: BorderRadius.circular(20),
-      ),
-      child: isBot
-          ? MarkdownBody(
-              data: widget.message.displayText,
-              imageBuilder: (uri, title, alt) => _buildImageWidget(uri.toString()),
-              styleSheet: MarkdownStyleSheet(
-                p: const TextStyle(
-                  fontSize: 15, 
-                  height: 1.5, 
-                  color: Color(0xFF000000),
-                  fontWeight: FontWeight.w400,
-                ),
-                code: TextStyle(
-                  backgroundColor: const Color(0xFFEAE9E5),
-                  color: const Color(0xFF000000),
-                  fontFamily: 'SF Mono',
-                  fontSize: 14,
-                ),
-                codeblockDecoration: BoxDecoration(
-                  color: const Color(0xFFEAE9E5),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                h1: const TextStyle(color: Color(0xFF000000), fontWeight: FontWeight.bold),
-                h2: const TextStyle(color: Color(0xFF000000), fontWeight: FontWeight.bold),
-                h3: const TextStyle(color: Color(0xFF000000), fontWeight: FontWeight.bold),
-                listBullet: const TextStyle(color: Color(0xFFA3A3A3)),
-                blockquote: const TextStyle(color: Color(0xFFA3A3A3)),
-                strong: const TextStyle(color: Color(0xFF000000), fontWeight: FontWeight.bold),
-                em: const TextStyle(color: Color(0xFF000000), fontStyle: FontStyle.italic),
-              ),
-            )
-          : Text(
-              widget.message.text, 
-              style: const TextStyle(
-                fontSize: 15, 
-                height: 1.5, 
-                color: Color(0xFF000000),
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-    );
-
-    return Align(
-      alignment: isBot ? Alignment.centerLeft : Alignment.centerRight,
-      child: Column(
-        crossAxisAlignment: isBot ? CrossAxisAlignment.start : CrossAxisAlignment.end,
-        children: [
-          // Thoughts panel for bot messages - MOVED ABOVE THE MESSAGE
-          if (isBot && widget.message.thoughts.isNotEmpty)
-            _ThoughtsPanel(thoughts: widget.message.thoughts),
-          if (isBot && widget.message.codes.isNotEmpty)
-            _CodePanel(codes: widget.message.codes),
-          // Tool results panel for bot messages
-          if (isBot && widget.message.toolData.isNotEmpty)
-            _ToolResultsPanel(toolData: widget.message.toolData),
-          // Agent processing panel removed – agent output will now stream directly in the chat bubble
-          if (isUser)
-            GestureDetector(
-              onTap: _toggleUserActions,
-              child: AnimatedContainer(
-                duration: const Duration(milliseconds: 200),
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(20),
-                  color: _showUserActions ? const Color(0xFFEAE9E5).withOpacity(0.3) : Colors.transparent,
-                ),
-                child: bubbleContent,
-              ),
-            )
-          else if (isBot && canShowActions)
-            GestureDetector(
-              onTap: _toggleActions,
-              child: AnimatedContainer(
-                duration: const Duration(milliseconds: 200),
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(20),
-                  color: _showActions ? const Color(0xFFEAE9E5).withOpacity(0.3) : Colors.transparent,
-                ),
-                child: bubbleContent,
-              ),
-            )
-          else
-            bubbleContent,
-          // User message actions
-          if (isUser && _showUserActions)
-            SlideTransition(
-              position: Tween<Offset>(
-                begin: const Offset(0.2, 0),
-                end: Offset.zero,
-              ).animate(_userActionsAnimation),
-              child: FadeTransition(
-                opacity: _userActionsAnimation,
-                child: Container(
-                  margin: const EdgeInsets.only(right: 12, top: 8, bottom: 8),
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      // Copy
-                      _ActionButton(
-                        icon: Icons.content_copy_rounded,
-                        onTap: () async {
-                          await Clipboard.setData(ClipboardData(text: widget.message.text));
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            SnackBar(
-                              content: const Text(
-                                '📋 Message copied to clipboard!',
-                                style: TextStyle(
-                                  color: Color(0xFF000000),
-                                  fontWeight: FontWeight.w500,
-                                ),
-                              ),
-                              backgroundColor: Colors.white,
-                              behavior: SnackBarBehavior.floating,
-                              shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-                              margin: const EdgeInsets.all(16),
-                              elevation: 4,
-                              duration: const Duration(seconds: 2),
-                            ),
-                          );
-                          _toggleUserActions();
-                        },
-                        tooltip: 'Copy text',
-                      ),
-                      const SizedBox(width: 8),
-                      // Edit & Resend
-                      _ActionButton(
-                        icon: Icons.edit_rounded,
-                        onTap: () {
-                          // Call the existing edit functionality
-                          if (widget.onUserMessageTap != null) {
-                            widget.onUserMessageTap!();
-                          }
-                          _toggleUserActions();
-                        },
-                        tooltip: 'Edit & Resend',
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          // iOS-style action buttons that slide in for bot messages
-          if (canShowActions && _showActions)
-            SlideTransition(
-              position: Tween<Offset>(
-                begin: const Offset(-0.2, 0),
-                end: Offset.zero,
-              ).animate(_actionsAnimation),
-              child: FadeTransition(
-                opacity: _actionsAnimation,
-                child: Container(
-                  margin: const EdgeInsets.only(left: 12, top: 8, bottom: 8),
-                  child: Row(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      // Thumbs up
-                      _ActionButton(
-                        icon: Icons.thumb_up_rounded,
-                        onTap: () => _giveFeedback(context, true),
-                        tooltip: 'Good response',
-                      ),
-                      const SizedBox(width: 8),
-                      // Thumbs down
-                      _ActionButton(
-                        icon: Icons.thumb_down_rounded,
-                        onTap: () => _giveFeedback(context, false),
-                        tooltip: 'Bad response',
-                      ),
-                      const SizedBox(width: 8),
-                      // Copy
-                      _ActionButton(
-                        icon: Icons.content_copy_rounded,
-                        onTap: () => _copyMessage(context),
-                        tooltip: 'Copy text',
-                      ),
-                      const SizedBox(width: 8),
-                      // Regenerate
-                      _ActionButton(
-                        icon: Icons.refresh_rounded,
-                        onTap: () {
-                          widget.onRegenerate?.call();
-                          _toggleActions();
-                        },
-                        tooltip: 'Regenerate',
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _ActionButton extends StatelessWidget {
-  final IconData icon;
-  final VoidCallback onTap;
-  final String? tooltip;
-  
-  const _ActionButton({
-    required this.icon, 
-    required this.onTap,
-    this.tooltip,
-  });
-  
-  @override
-  Widget build(BuildContext context) {
-    return Tooltip(
-      message: tooltip ?? '',
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(8),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 4,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      textStyle: const TextStyle(
-        color: Colors.black,
-        fontSize: 12,
-        fontWeight: FontWeight.w500,
-      ),
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          onTap: () {
-            HapticFeedback.lightImpact();
-            onTap();
-          },
-          borderRadius: BorderRadius.circular(16),
-          child: Container(
-            padding: const EdgeInsets.all(10),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(16),
-              border: Border.all(
-                color: const Color(0xFFE0E0E0),
-                width: 1,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.08),
-                  blurRadius: 4,
-                  offset: const Offset(0, 2),
-                ),
-              ],
-            ),
-            child: Icon(
-              icon, 
-              color: const Color(0xFF000000), 
-              size: 16,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
 
 /* ----------------------------------------------------------
-   INPUT BAR – Clean Design with Icons Below
+   END OF CHAT PAGE IMPLEMENTATION
 ---------------------------------------------------------- */
-class _InputBar extends StatelessWidget {
-  const _InputBar({
-    required this.controller,
-    required this.onSend,
-    required this.onStop,
-    required this.awaitingReply,
-    required this.isEditing,
-    required this.onCancelEdit,
-    required this.externalToolsService,
-    required this.webSearchMode,
-    required this.onToggleWebSearch,
-    required this.onImageUpload,
-    this.uploadedImagePath,
-    required this.onClearImage,
-  });
-
-  final TextEditingController controller;
-  final VoidCallback onSend;
-  final VoidCallback onStop;
-  final bool awaitingReply;
-  final bool isEditing;
-  final VoidCallback onCancelEdit;
-  final ExternalToolsService externalToolsService;
-  final bool webSearchMode;
-  final VoidCallback onToggleWebSearch;
-  final VoidCallback onImageUpload;
-  final String? uploadedImagePath;
-  final VoidCallback onClearImage;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.only(bottom: 16),
-      decoration: const BoxDecoration(
-        color: Color(0xFFF4F3F0), // Main theme background
-      ),
-      child: Column(
-        children: [
-          // Edit mode indicator
-          if (isEditing)
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 12),
-              margin: const EdgeInsets.only(left: 20, right: 20, bottom: 12, top: 16),
-              decoration: BoxDecoration(
-                color: const Color(0xFF000000).withOpacity(0.1),
-                borderRadius: BorderRadius.circular(12),
-                border: Border.all(
-                  color: const Color(0xFF000000).withOpacity(0.2),
-                ),
-              ),
-              child: Row(
-                children: [
-                  const Icon(Icons.edit_rounded, color: Color(0xFF000000), size: 20),
-                  const SizedBox(width: 12),
-                  const Expanded(
-                    child: Text(
-                      "Editing message...", 
-                      style: TextStyle(color: Color(0xFF000000), fontWeight: FontWeight.w500),
-                    ),
-                  ),
-                  GestureDetector(
-                    onTap: () {
-                      HapticFeedback.lightImpact();
-                      onCancelEdit();
-                    },
-                    child: Container(
-                      padding: const EdgeInsets.all(4),
-                      decoration: BoxDecoration(
-                        color: const Color(0xFF000000),
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      child: const Icon(Icons.close_rounded, color: Colors.white, size: 16),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          
-          // Main input container (smaller height)
-          Container(
-            margin: EdgeInsets.fromLTRB(20, isEditing ? 0 : 16, 20, 0),
-            clipBehavior: Clip.antiAlias,
-            decoration: BoxDecoration(
-              color: Colors.white, // White input background
-              borderRadius: BorderRadius.circular(24), // Fully rounded border on both sides
-              border: Border.all(
-                color: const Color(0xFFEAE9E5),
-                width: 1,
-              ),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.08),
-                  blurRadius: 12,
-                  offset: const Offset(0, 4),
-                ),
-              ],
-            ),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.end,
-              children: [
-                // Text input field with reduced height
-                Expanded(
-                  child: TextField(
-                    controller: controller,
-                    enabled: !awaitingReply,
-                    maxLines: 3, // Reduced from 6
-                    minLines: 1, // Reduced from 3
-                    textCapitalization: TextCapitalization.sentences,
-                    cursorColor: const Color(0xFF000000),
-                    textInputAction: TextInputAction.send,
-                    onSubmitted: (_) => onSend(),
-                    style: const TextStyle(
-                      color: Color(0xFF000000),
-                      fontSize: 16,
-                      height: 1.4,
-                    ),
-                    decoration: InputDecoration(
-                      hintText: awaitingReply 
-                          ? 'AhamAI is responding...' 
-                          : externalToolsService.isExecuting
-                              ? 'External tool is running...'
-                              : webSearchMode
-                                  ? 'Web search mode - Ask me anything...'
-                                  : uploadedImagePath != null
-                                      ? 'Image uploaded - Describe or ask about it...'
-                                      : 'Message AhamAI (images, web search, screenshots, vision)...',
-                      hintStyle: const TextStyle(
-                        color: Color(0xFFA3A3A3),
-                        fontSize: 16,
-                        height: 1.4,
-                      ),
-                      border: InputBorder.none,
-                      enabledBorder: InputBorder.none,
-                      focusedBorder: InputBorder.none,
-                      errorBorder: InputBorder.none,
-                      focusedErrorBorder: InputBorder.none,
-                      disabledBorder: InputBorder.none,
-                      contentPadding: const EdgeInsets.symmetric(
-                        horizontal: 20, // Increased padding for better rounded appearance
-                        vertical: 12 // Reduced from 18
-                      ),
-                    ),
-                  ),
-                ),
-                
-                // Send/Stop button
-                Padding(
-                  padding: const EdgeInsets.only(right: 12, bottom: 6), // Adjusted padding
-                  child: GestureDetector(
-                    onTap: awaitingReply ? onStop : onSend,
-                    child: AnimatedContainer(
-                      duration: const Duration(milliseconds: 200),
-                      padding: const EdgeInsets.all(10), // Smaller padding
-                      decoration: BoxDecoration(
-                        color: awaitingReply 
-                            ? Colors.red.withOpacity(0.1)
-                            : const Color(0xFF000000),
-                        borderRadius: BorderRadius.circular(12), // Smaller radius
-                      ),
-                      child: Icon(
-                        awaitingReply ? Icons.stop_circle : Icons.arrow_upward_rounded,
-                        color: awaitingReply ? Colors.red : Colors.white,
-                        size: 18, // Smaller icon
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          
-          // Icons below input bar
-          if (!awaitingReply)
-            Container(
-              margin: const EdgeInsets.fromLTRB(20, 12, 20, 0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  // Image preview removed as requested
-                  
-                  // Action icons row
-                  Row(
-                    children: [
-
-                        
-                                                 // Web Search Icon - clean design
-                         _AnimatedModeIcon(
-                           isActive: webSearchMode,
-                           icon: FontAwesomeIcons.search,
-                           label: 'Search',
-                          onTap: () {
-                            HapticFeedback.lightImpact();
-                            onToggleWebSearch();
-                          },
-                        ),
-                        
-                        const SizedBox(width: 12),
-                        
-                        // Image Upload Icon - clean design
-                        _AnimatedModeIcon(
-                          isActive: uploadedImagePath != null,
-                          icon: uploadedImagePath != null 
-                              ? FontAwesomeIcons.times
-                              : FontAwesomeIcons.camera,
-                          label: 'Image',
-                          onTap: () {
-                            HapticFeedback.lightImpact();
-                            if (uploadedImagePath != null) {
-                              onClearImage();
-                            } else {
-                              onImageUpload();
-                            }
-                          },
-                        ),
-                      
-                      const Spacer(),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _ShareButton extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final VoidCallback onTap;
-
-  const _ShareButton({
-    required this.icon,
-    required this.label,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        HapticFeedback.lightImpact();
-        onTap();
-      },
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Container(
-            padding: const EdgeInsets.all(12),
-            decoration: BoxDecoration(
-              color: const Color(0xFF2C2C2E),
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Icon(
-              icon,
-              color: const Color(0xFF4A9B8E),
-              size: 24,
-            ),
-          ),
-          const SizedBox(height: 8),
-          Text(
-            label,
-            style: const TextStyle(
-              color: Color(0xFF8E8E93),
-              fontSize: 12,
-              fontWeight: FontWeight.w500,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/* ----------------------------------------------------------
-   CODE PANEL - Collapsible panel for code blocks
----------------------------------------------------------- */
-class _CodePanel extends StatefulWidget {
-  final List<CodeContent> codes;
-  
-  const _CodePanel({required this.codes});
-  
-  @override
-  State<_CodePanel> createState() => _CodePanelState();
-}
-
-class _CodePanelState extends State<_CodePanel> with SingleTickerProviderStateMixin {
-  bool _isExpanded = false;
-  late AnimationController _animationController;
-  late Animation<double> _expandAnimation;
-  
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
-    _expandAnimation = CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.easeInOut,
-    );
-  }
-  
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-  
-  void _toggleExpansion() {
-    setState(() {
-      _isExpanded = !_isExpanded;
-      if (_isExpanded) {
-        _animationController.forward();
-      } else {
-        _animationController.reverse();
-      }
-    });
-  }
-  
-  void _copyCode(String code) {
-    Clipboard.setData(ClipboardData(text: code));
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: const Text(
-          '📋 Code copied to clipboard!',
-          style: TextStyle(
-            color: Color(0xFF000000),
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-        margin: const EdgeInsets.all(16),
-        elevation: 4,
-        duration: const Duration(seconds: 2),
-      ),
-    );
-  }
-  
-  String _getCodePreview() {
-    if (widget.codes.isEmpty) return '';
-    final firstCode = widget.codes.first;
-    final preview = firstCode.code.length > 50 
-        ? '${firstCode.code.substring(0, 50)}...'
-        : firstCode.code;
-    return preview;
-  }
-  
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(top: 8),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF4F3F0),
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Header with toggle button
-          InkWell(
-            onTap: _toggleExpansion,
-            borderRadius: BorderRadius.circular(12),
-            child: Padding(
-              padding: const EdgeInsets.all(12),
-              child: Row(
-                children: [
-                  Icon(
-                    Icons.code_rounded,
-                    size: 16,
-                    color: const Color(0xFF000000),
-                  ),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      _isExpanded 
-                          ? '${widget.codes.length} Code Block${widget.codes.length > 1 ? 's' : ''}'
-                          : _getCodePreview(),
-                      style: GoogleFonts.inter(
-                        fontSize: 13,
-                        color: const Color(0xFFA3A3A3),
-                        fontWeight: FontWeight.w500,
-                      ),
-                      maxLines: _isExpanded ? 1 : 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                  AnimatedRotation(
-                    turns: _isExpanded ? 0.5 : 0,
-                    duration: const Duration(milliseconds: 300),
-                    child: Icon(
-                      Icons.keyboard_arrow_down_rounded,
-                      size: 18,
-                      color: const Color(0xFF000000),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          // Expandable content
-          SizeTransition(
-            sizeFactor: _expandAnimation,
-            child: Container(
-              padding: const EdgeInsets.fromLTRB(12, 0, 12, 12),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: widget.codes.map((codeContent) => Container(
-                  margin: const EdgeInsets.only(bottom: 8),
-                  decoration: BoxDecoration(
-                    color: const Color(0xFFEAE9E5),
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // Header with language and copy button
-                      Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                        child: Row(
-                          children: [
-                            Container(
-                              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                              decoration: BoxDecoration(
-                                color: const Color(0xFF000000),
-                                borderRadius: BorderRadius.circular(12),
-                              ),
-                              child: Text(
-                                codeContent.extension.toUpperCase(),
-                                style: GoogleFonts.inter(
-                                  fontSize: 10,
-                                  fontWeight: FontWeight.w600,
-                                  color: const Color(0xFFFFFFFF),
-                                  letterSpacing: 0.5,
-                                ),
-                              ),
-                            ),
-                            const Spacer(),
-                            IconButton(
-                              onPressed: () => _copyCode(codeContent.code),
-                              icon: const Icon(
-                                Icons.copy_rounded,
-                                size: 16,
-                                color: Color(0xFF000000),
-                              ),
-                              constraints: const BoxConstraints(
-                                minWidth: 32,
-                                minHeight: 32,
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      // Code content
-                      Container(
-                        width: double.infinity,
-                        padding: const EdgeInsets.fromLTRB(12, 0, 12, 12),
-                        child: Container(
-                          padding: const EdgeInsets.all(12),
-                          decoration: BoxDecoration(
-                            color: Colors.black87,
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                          child: SingleChildScrollView(
-                            scrollDirection: Axis.horizontal,
-                            child: Text(
-                              codeContent.code,
-                              style: const TextStyle(
-                                fontFamily: 'monospace',
-                                fontSize: 12,
-                                height: 1.4,
-                                color: Color(0xFFFFFFFF),
-                                fontWeight: FontWeight.w400,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                )).toList(),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/* ----------------------------------------------------------
-   THOUGHTS PANEL - Collapsible panel for AI thinking content
----------------------------------------------------------- */
-class _ThoughtsPanel extends StatefulWidget {
-  final List<ThoughtContent> thoughts;
-  
-  const _ThoughtsPanel({required this.thoughts});
-  
-  @override
-  State<_ThoughtsPanel> createState() => _ThoughtsPanelState();
-}
-
-class _ThoughtsPanelState extends State<_ThoughtsPanel> with SingleTickerProviderStateMixin {
-  bool _isExpanded = false;
-  late AnimationController _animationController;
-  late Animation<double> _expandAnimation;
-  
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
-    _expandAnimation = CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.easeInOut,
-    );
-  }
-  
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-  
-  void _toggleExpansion() {
-    setState(() {
-      _isExpanded = !_isExpanded;
-      if (_isExpanded) {
-        _animationController.forward();
-      } else {
-        _animationController.reverse();
-      }
-    });
-  }
-  
-  String _getThoughtsPreview() {
-    if (widget.thoughts.isEmpty) return '';
-    final firstThought = widget.thoughts.first;
-    final preview = firstThought.text.length > 50 
-        ? '${firstThought.text.substring(0, 50)}...'
-        : firstThought.text;
-    return preview;
-  }
-  
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(left: 12, top: 8, bottom: 4),
-      constraints: const BoxConstraints(maxWidth: 320),
-      decoration: BoxDecoration(
-        color: const Color(0xFFEAE9E5),
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Header with toggle button
-          InkWell(
-            onTap: _toggleExpansion,
-            borderRadius: BorderRadius.circular(12),
-            child: Padding(
-              padding: const EdgeInsets.all(12),
-              child: Row(
-                children: [
-                  Icon(
-                    Icons.psychology_rounded,
-                    size: 16,
-                    color: const Color(0xFF000000),
-                  ),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      _isExpanded 
-                          ? 'Thoughts (${widget.thoughts.length})'
-                          : _getThoughtsPreview(),
-                      style: const TextStyle(
-                        fontSize: 13,
-                        color: Color(0xFFA3A3A3),
-                        fontWeight: FontWeight.w500,
-                      ),
-                      maxLines: _isExpanded ? 1 : 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                  AnimatedRotation(
-                    turns: _isExpanded ? 0.5 : 0,
-                    duration: const Duration(milliseconds: 300),
-                    child: Icon(
-                      Icons.keyboard_arrow_down_rounded,
-                      size: 18,
-                      color: const Color(0xFF000000),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          // Expandable content
-          SizeTransition(
-            sizeFactor: _expandAnimation,
-            child: Container(
-              padding: const EdgeInsets.fromLTRB(12, 0, 12, 12),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: widget.thoughts.map((thought) => Container(
-                  margin: const EdgeInsets.only(bottom: 8),
-                  padding: const EdgeInsets.all(12),
-                  decoration: BoxDecoration(
-                    color: const Color(0xFFF4F3F0),
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                        decoration: BoxDecoration(
-                          color: const Color(0xFF000000),
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        child: Text(
-                          thought.type.toUpperCase(),
-                          style: const TextStyle(
-                            fontSize: 10,
-                            fontWeight: FontWeight.w600,
-                            color: Color(0xFFFFFFFF),
-                            letterSpacing: 0.5,
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        thought.text,
-                        style: const TextStyle(
-                          fontSize: 13,
-                          height: 1.4,
-                          color: Color(0xFF000000),
-                          fontWeight: FontWeight.w400,
-                        ),
-                      ),
-                    ],
-                  ),
-                )).toList(),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/* ----------------------------------------------------------
-   TOOL RESULTS PANEL - Displays external tool execution results
----------------------------------------------------------- */
-class _ToolResultsPanel extends StatefulWidget {
-  final Map<String, dynamic> toolData;
-  
-  const _ToolResultsPanel({required this.toolData});
-  
-  @override
-  State<_ToolResultsPanel> createState() => _ToolResultsPanelState();
-}
-
-class _ToolResultsPanelState extends State<_ToolResultsPanel> with SingleTickerProviderStateMixin {
-  bool _isExpanded = true; // Start expanded for better visibility
-  late AnimationController _animationController;
-  late Animation<double> _expandAnimation;
-  
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
-    _expandAnimation = CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.easeInOut,
-    );
-    // Start expanded
-    _animationController.forward();
-  }
-  
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-  
-  void _toggleExpansion() {
-    setState(() {
-      _isExpanded = !_isExpanded;
-      if (_isExpanded) {
-        _animationController.forward();
-      } else {
-        _animationController.reverse();
-      }
-    });
-  }
-  
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(top: 8),
-      decoration: BoxDecoration(
-        color: const Color(0xFFF4F3F0),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: const Color(0xFF000000).withOpacity(0.1)),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Header with toggle button
-          InkWell(
-            onTap: _toggleExpansion,
-            borderRadius: BorderRadius.circular(12),
-            child: Padding(
-              padding: const EdgeInsets.all(12),
-              child: Row(
-                children: [
-                  Icon(
-                    Icons.build_circle_rounded,
-                    size: 16,
-                    color: const Color(0xFF000000),
-                  ),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      'Tool Results (${widget.toolData.length})',
-                      style: GoogleFonts.inter(
-                        fontSize: 13,
-                        color: const Color(0xFF000000),
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                  AnimatedRotation(
-                    turns: _isExpanded ? 0.5 : 0,
-                    duration: const Duration(milliseconds: 300),
-                    child: Icon(
-                      Icons.keyboard_arrow_down_rounded,
-                      size: 18,
-                      color: const Color(0xFF000000),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          // Expandable content
-          SizeTransition(
-            sizeFactor: _expandAnimation,
-            child: Container(
-              padding: const EdgeInsets.fromLTRB(12, 0, 12, 12),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: widget.toolData.entries.map((entry) {
-                  final toolName = entry.key;
-                  final result = entry.value as Map<String, dynamic>;
-                  
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 8),
-                    decoration: BoxDecoration(
-                      color: const Color(0xFFEAE9E5),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        // Header with tool name and status
-                        Container(
-                          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                          child: Row(
-                            children: [
-                              Container(
-                                padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                                decoration: BoxDecoration(
-                                  color: result['success'] == true ? Colors.green : Colors.red,
-                                  borderRadius: BorderRadius.circular(12),
-                                ),
-                                child: Text(
-                                  toolName.toUpperCase(),
-                                  style: GoogleFonts.inter(
-                                    fontSize: 10,
-                                    fontWeight: FontWeight.w600,
-                                    color: const Color(0xFFFFFFFF),
-                                    letterSpacing: 0.5,
-                                  ),
-                                ),
-                              ),
-                              const SizedBox(width: 8),
-                              Icon(
-                                result['success'] == true ? Icons.check_circle : Icons.error,
-                                size: 16,
-                                color: result['success'] == true ? Colors.green : Colors.red,
-                              ),
-                              const Spacer(),
-                              if (result['execution_time'] != null)
-                                Text(
-                                  'Executed',
-                                  style: GoogleFonts.inter(
-                                    fontSize: 10,
-                                    color: const Color(0xFFA3A3A3),
-                                  ),
-                                ),
-                            ],
-                          ),
-                        ),
-                        // Tool result content
-                        Container(
-                          width: double.infinity,
-                          padding: const EdgeInsets.fromLTRB(12, 0, 12, 12),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              
-                              // Result details
-                              if ((result['url'] != null) ||
-                                  (result['models'] != null) ||
-                                  (result['new_model'] != null) ||
-                                  (result['api_status'] != null) ||
-                                  (result['error'] != null))
-                                Container(
-                                  padding: const EdgeInsets.all(12),
-                                  decoration: BoxDecoration(
-                                    color: Colors.black87,
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                  child: Column(
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: [
-                                      if (result['url'] != null)
-                                        _buildResultRow('URL', result['url']),
-                                      if (result['models'] != null)
-                                        _buildResultRow('Models Count', '${(result['models'] as List).length}'),
-                                      if (result['new_model'] != null)
-                                        _buildResultRow('New Model', result['new_model']),
-                                      if (result['api_status'] != null)
-                                        _buildResultRow('API Status', result['api_status']),
-                                      if (result['error'] != null)
-                                        _buildResultRow('Error', result['error'], isError: true),
-                                    ],
-                                  ),
-                                ),
-
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  );
-                }).toList(),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-  
-  Widget _buildResultRow(String label, String value, {bool isError = false}) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 4),
-      child: RichText(
-        text: TextSpan(
-          children: [
-            TextSpan(
-              text: '$label: ',
-              style: TextStyle(
-                fontFamily: 'monospace',
-                fontSize: 12,
-                height: 1.4,
-                color: Colors.grey[400],
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            TextSpan(
-              text: value,
-              style: TextStyle(
-                fontFamily: 'monospace',
-                fontSize: 12,
-                height: 1.4,
-                color: isError ? Colors.red[300] : const Color(0xFFFFFFFF),
-                fontWeight: FontWeight.w400,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-  
-}
-
-/* ----------------------------------------------------------
-   ANIMATED MODE ICON - Reusable component with animated border
----------------------------------------------------------- */
-class _AnimatedModeIcon extends StatefulWidget {
-  final bool isActive;
-  final IconData icon;
-  final String label;
-  final VoidCallback onTap;
-
-  const _AnimatedModeIcon({
-    required this.isActive,
-    required this.icon,
-    required this.label,
-    required this.onTap,
-  });
-
-  @override
-  State<_AnimatedModeIcon> createState() => _AnimatedModeIconState();
-}
-
-class _AnimatedModeIconState extends State<_AnimatedModeIcon> 
-    with SingleTickerProviderStateMixin {
-  
-  late AnimationController _animationController;
-  late Animation<double> _scaleAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 150),
-      vsync: this,
-    );
-    _scaleAnimation = Tween<double>(
-      begin: 1.0,
-      end: 0.95,
-    ).animate(CurvedAnimation(
-      parent: _animationController,
-      curve: Curves.easeInOut,
-    ));
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTapDown: (_) => _animationController.forward(),
-      onTapUp: (_) => _animationController.reverse(),
-      onTapCancel: () => _animationController.reverse(),
-      onTap: widget.onTap,
-      child: AnimatedBuilder(
-        animation: _scaleAnimation,
-        builder: (context, child) {
-          return Transform.scale(
-            scale: _scaleAnimation.value,
-            child: Container(
-              padding: const EdgeInsets.all(8),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  // Clean icon with subtle active state
-                  Container(
-                    width: 40,
-                    height: 40,
-                    decoration: BoxDecoration(
-                      color: widget.isActive 
-                          ? const Color(0xFF6366F1).withOpacity(0.1)
-                          : Colors.transparent,
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    child: Center(
-                      child: FaIcon(
-                        widget.icon,
-                        color: widget.isActive 
-                            ? const Color(0xFF6366F1)
-                            : const Color(0xFF6B7280),
-                        size: 18,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(
-                    widget.label,
-                    style: GoogleFonts.inter(
-                      fontSize: 10,
-                      fontWeight: widget.isActive ? FontWeight.w600 : FontWeight.w500,
-                      color: widget.isActive 
-                          ? const Color(0xFF6366F1)
-                          : const Color(0xFF6B7280),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
+// All components have been moved to separate files:
+// - chat_message_bubble.dart
+// - chat_input_bar.dart 
+// - chat_panels.dart
+// - tool_loading_indicator.dart
 }
